overlays: Fixed dtc warnings
Fixed dtc warnings, and suppressed the ones which have no sense in
the context:

-Wno-unit_address_vs_reg : stops the warnings for having a child node without
a "reg". In our its files, we have hashes as child nodes, without a reg
property.
-Wno-graph_child_address : stops the warning that says to not use cells
for nodes with a single child. If we remove the cells information, we get
another warning: that cells need to be added.
-Wno-pwms_property: pwm cells property is inherited from base dtb. So the
warning doesn't make sense for the overlays here.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>