sama5d2_xplained_grts: add its and overlay for resistive touchscreen board
This overlay is done for an adapted board based on SAMA5D2_Xplained.
It has the PMIC moved to another i2c bus, so we can use the ADC with the
resistive touchscreen.
Another modified board is required to adapt the resistive touchscreen
SAM9X5-DM to a ribbon cable that can be connected to the LCD port of the
Xplained board.
This board is for demonstration purposes, it proves the functionality
of the resistive touchscreen and ADC of the SoC.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>