sam9x60: wilc: add phandle to pwrseq driver
Add phandle to WILC power sequence driver. Otherwise it will not
be probed by MMC subsystem.

Note: There is a known issue that can occur on the SAM9X60-EK Rev. B,
Rev. 2 and Rev. 3 boards if the user tries tomount and use the optional
WILC3000 module together with the LCD or with the Ethernet PHY.

With WILC3000, a special power-up sequence applies, requiring that the
reset line be held low for a specific period of time. However, as all devices
share the same reset line, this can cause communication disruptions during
runtime.

To solve this issue, we recommend separating the reset line going to the
WILC3000 by cutting the trace and connecting it to a free GPIO from the 40-pin
GPIO connector (J16). Our recommendation is to use PC31 (pin 29).

For more details consult "Powering-up Issue" chapter from SAM9X60-EK User's
Guide Errata.

Fixes: e20700d2dad0 ("sam9x60ek: add support for WILC 3000")
Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>
[cristian.birsan@microchip.com add power-up issue note]
Acked-by : Cristian Birsan <cristian.birsan@microchip.com>