sama7g5ek: isc_imx219: add endpoints connection between csi2dc and csi2host
Connection between csi2dc and csi2host has to be at the overlay level.
If a parallel sensor is used, it will be connected to the csi2dc.
Thus, the connection between csi2dc and csi2host cannot be permanent, must
be selected only in the case of a serial mipi csi2 sensor.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
Acked-by: Cristian Birsan <cristian.birsan@microchip.com>