treewide: replace '@' with '-' for hash nodes
According to upstream u-boot-tools commit :

Fixes: 79af75f777 ("fit: Don't allow verification of images with @ nodes")

"When searching for a node called 'fred', any unit address appended to the
name is ignored by libfdt, meaning that 'fred' can match 'fred@1'. This
means that we cannot be sure that the node originally intended is the one
that is used.

Disallow use of nodes with unit addresses."

Replace all '@' with dashes in the nodes.

Without this change, when building with newer mkimage (>2021.04):

mkimage: verify_header failed for FIT Image support with exit code 1

Thus, mkimage fails to build the ITBs.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
Reported-by: Codrin Ciubotariu <codrin.ciubotariu@microchip.com>
Reported-by: Mihai Sain <mihai.sain@microchip.com>
Acked-by: Cristian Birsan <cristian.birsan@microchip.com>