sam9x60ek: wilc: enable pwrseq driver
The initial role of this overlay was to have it as reference for
customers and use it after the PCB has been properly modified.
In the initial version of this commit both sdmmc1 and wifi_pwrseq
nodes were disabled.

Starting with commit 948e0c2cd14e ("sam9x60ek: wilc: Enable sdmmc1
in overlay") sdmmc1 node is enabled. Even so, the sdmmc driver for
sdmmc1 node will not be probed as when the mmc-pwrseq phandle is
provided in DT the sdmmc code is parsing it and postpone the sdmmc
driver probe in case the pwrseq driver provided via mmc-pwrseq
phandle is not probed.

With all these enable also the pwrseq driver in device tree to
avoid the above scenario. In case client will change the board
the overlay will not need to be recompiled but just loaded.

Reported-by: Xing Chen <xing.chen@microchip.com>
Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>