sama7g5ek: isc_imx*: add flx8 into TWI mode
The Flexcom8 is already into TWI mode by board DTS.
However this overlay blob is used by people as an example on how to
add an overlay for a sensor, and the sensor can be moved to a different i2c
bus. Thus, for consistency add the TWI mode property to the parent flx node.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
Acked-by: Cristian Birsan <cristian.birsan@microchip.com>