mpfs: fix paths and addresses
Buildroot External expects paths and names of a certain
format, but the current its is a copy and paste of the its
in the Buildroot SDK. Update the paths to support building
an itb using Buildroot External.
The Buildroot External kernel uses a builtin initramfs
and therefore requires more space. Update the load address
of the device tree blob to accomodate this change.
Update default configuration and remove old
initramfs reference as a consequence of the above changes.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>