Add dt-overlay validation script
In order to validate Devicetree Overlays, add a Bash script that makes
use of the Devicetree bindings located in the Linux Kernel tree to check
each overlay against every relevant Devicetree binding that has been
converted to json-schema. In order to validate each Devicetree Overlay of
a given board, the script does the following:
1. Depending on the directory with the Deviectree Overlays, the script
sets up some environment variables regarding the location of the DT files
from the Linux Kernel tree (DTSO_DIR_TO_DT variable).
2. It then runs the defconfig linked to the board that owns the overlays
(DTSO_DIR_TO_CFG variable).
3. The `dt_binding_check` recipe of the Linux Kernel's build system is
then run to obtain the final processed-schema.yaml.
4. For each Devicetree Overlay contained within the given directory:
	4.1. Dumps the overlay into the board's DTS file.
	4.2. Re-builds the DTB of the board.
	4.3. Validates the DTB against the processed-schema.json
	     generated at step 3.
5. Restore the DTS file.

The script, by default, filters out any output related to the build
process, except relevant information such as errors or warnings.

The DTS file will be restored upon either the end of the execution or
the occurrence of an interrupt caused by Ctrl-C to the default
Linux Kernel tree directory (which is the user's KERNEL_DIR environment
variable, but may be changed with the help of `-k` option as well).

A helper `-h` command line option can be issued to see all of the
currently available options.

An example on how to run the script, supposing we are in the
dt-overlay directory, would be:

$ ./scripts/dt_overlay_check.sh -b <board-directory> -v

In the above example, the argument passed to the `-b` option is
the path to one of the board directories containing the
dtso files (e.g. sam9x60ek, sama5d27_som1_ek, sama5d27_wlsom1_ek,
etc.). The option `-v`, as the output of `-h` says, is optional and
simply enables some green colored debug messages that could tell the
user at which step the script is and what it is actually currently doing.

Furthermore, add a Makefile recipe for checking all available DTSO files.

Signed-off-by: Sergiu Moga <sergiu.moga@microchip.com>
Reviewed-by: Conor Dooley <conor.dooley@microchip.com>
Acked-by: Cristian Birsan <cristian.birsan@microchip.com>