scripts: dt_overlay_check: Be aware of relative/empty path in `KERNEL_DIR`
The script does not take into consideration the possibility of having the
implicit `KERNEL_DIR` environment variable as a relative path, although
the handling of the `-k` command line option does. So use the `realpath`
command on the implicit `KERNEL_DIR` environment variable as well to
ensure that this possibility is taken care of and ignore its error output
since we later check its result anyway.

Furthermore, the KERNEL_DIR variable might also be empty, in which case
the user should know about this, instead of just printing nothing and
erroring out. So also check whether KERNEL_DIR is empty or not.

Signed-off-by: Sergiu Moga <sergiu.moga@microchip.com>
Acked-by: Cristian Birsan <cristian.birsan@microchip.com>