sama5d3-eds: ksz9477: SPI speed and cleanup
Changes here include increasing the SPI bus speed to 20MHz.
While the EVB-KSZ9477 will support 33MHz, there is currently
an issue with the atmel SPI driver that is preventing this.
The board will run at 33MHz under the v5.15.67 kernel.

The RGMII mode is changed to reflect the default KSZ9477
rgmii mode and to make this ksz9477 example more clear with
regards to the phy-mode and delay properties. The 2-msec delay
can be placed on either the MAC or the PHY side. Here, the
delay is on the transmitter side for both devices.

Also cleaned up some of the comments.

Signed-off-by: Jerry Ray <jerry.ray@microchip.com>
Acked-by: Cristian Birsan <cristian.birsan@microchip.com>