mpfs_icicle: add sample AMP device tree as an dt overlay
Add a sample dtso file for the Icicle Kit in Asymmetric Multiprocessing
mode. Enable the Inter-Hart Communication subsystem (IHC), as well as
remoteproc for communication and life cycle management of remote
contexts. Finally, disable hart4 and mmuart3 since they are in use by
the PolarFire SoC AMP example demos running in the other context.

Drop the mpfs_icicle_amp directory in this repository, since it is no
longer needed.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: Cristian Birsan <cristian.birsan@microchip.com>