scripts: dt_overlay_check: fix dirty kernel sourcedir when not used as builddir
When the kernel source directory is not used as the build directory, the
script dirties the sourcedir with the .config it creates, because the
defconfig make command is reset after the scripts arguments are
converted into make arguments and appended to the commands.

Set the defconfig make command before this happens to fix the build.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>