mpfs_icicle: use microchip,ipc-sbi-remoteproc compatible in AMP dt overlay
Update the Icicle AMP device tree overlay to use the
"microchip,ipc-sbi-remoteproc" compatible string instead of
microchip,miv-remoteproc. This is because the Icicle Kit Reference
design release v2025.07 will use the the the MI-V IHC IP v2
available in the Libero catalogue.

The Mi-V IHC IP version 2 is not backwards compatible with the
Mi-V IHC subsystem used in the Icicle Kit reference design 2025.03
or earlier. For this reason, the AMP support in the v2025.07 release
will not be backwards compatible with previous releases.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Tested-by: Jamie Gibbons <jamie.gibbons@microchip.com>