sama7g5ek: Add support for WILCS02 mmc_spi module on mikroBUS1
This patch adds support of wilcs02 mmc-spi module on the sama7g5ek
host's mikroBUS1 slot. The support for this WILCS02 module is configured
as a seperate device tree overlay #wilc_mmc_spi in order to prevent
driver conflicts with WILCS02 SDIO module, which both uses the same
wilc-sdio.ko driver. Also, CONFIG_MMC_SPI must be enabled in the kernel
to enable support of this module on the host.

Signed-off-by: Prasurjya Rohan Saikia <Prasurjya.Rohansaikia@microchip.com>
Acked-by: Cristian Birsan <cristian.birsan@microchip.com>